Adjusted positioning of opt notation in grammar.
The tightness of "opt" relative to the grammar element that it marks varies depending upon whether it is marking code, which is styled with padding, or the name of a grammar part, which is not styled with padding.  We could style the names of grammar parts to add padding, but it was easier to simply add a class (grammar sub.loosen) to override the normal behavior of (grammar sub).