Fix for tighness of opt annotation; added comments.
In some cases, the "opt" grammar annotation is too tight to item it modifies.  As a temporary fix, adding a class (grammar sub.loosen) that loosens the spacing.

Also, added numerous comments to explain the purpose of each style and to note which styles  are to be moved to new files.